Refactor: move more test files into sub-packages
