+++ tests/test_misc/test_config.py
