+++ tests/test_misc/test_logging_setup.py
@@ -1,13 +1,12 @@
 import argparse
 import logging
 import logging.handlers
-from tests.test_logging_config import get_exp_formatted_time
+from tests.test_misc.test_logging_config import get_exp_formatted_time
 from typing import Optional
 
 from config import Paths
 from config import ScadaSettings
 from config import DEFAULT_LOG_FILE_NAME
-from config import LoggerLevels
 from config import LoggingSettings
 from config import RotatingFileHandlerSettings
 from logging_setup import setup_logging
