+++ tests/test_misc/test_scripts.py
