+++ tests/test_proactor/test_persister.py
@@ -10,7 +10,6 @@ from gwproto.messages import ProblemEvent
 from result import Result
 
 from config import ScadaSettings
-from proactor.persister import _PersistedItem
 from proactor.persister import FileExistedWarning
 from proactor.persister import FileMissing
 from proactor.persister import FileMissingWarning
